requirements.txt
fastapi>=0.95.0
uvicorn>=0.20.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0
langchain-groq>=0.1.0
langchain-core>=0.1.0
streamlit>=1.22.0
streamlit-extras>=0.2.0
pytest>=7.0.0
pytest-asyncio>=0.20.0
httpx>=0.23.0
requests>=2.28.0